requirements.txt
dateparser==1.0.0
pandas==1.3.2
google_auth_oauthlib==0.4.6
matplotlib==3.4.3
google_api_python_client==2.19.1
protobuf==3.17.3
